+++ objectstore/ArchiveQueueShard.hpp
@@ -47,7 +47,7 @@ public:
   struct JobInfo {
     uint64_t size;
     std::string address;
-    uint16_t copyNb;
+    uint32_t copyNb;
     uint64_t priority;
     uint64_t minArchiveRequestAge;
     uint64_t maxDrivesAllowed;
